requirements.txt
alembic==1.15.2
bcrypt==4.0.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
concurrent-log-handler==0.9.25
cryptography==3.4.7
Flask==3.1.0
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
greenlet==3.2.0
idna==3.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
mysql-connector-python==9.3.0
portalocker==3.1.1
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dotenv==1.1.0
pywin32==310
requests==2.32.3
SQLAlchemy==2.0.40
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
zipp==3.21.0
